changed tempo to float so decimal bpm allowed

+++ Core/Inc/lui_piezo_music.h
@@ -13,8 +13,8 @@ void piezo_init(void);
 float get_freq_from_note(char note[]);
 void pause_pwm(int duration_in_ms);
 void play_frequency(int duration_ms, float frequency);
-void play_frequency_array(float beats[], float frequencies[], int tempo, int song_len);
-void play_tune(float beats[], char *song_notes, int tempo);
+void play_frequency_array(float beats[], float frequencies[], float tempo, int song_len);
+void play_tune(float beats[], char *song_notes, float tempo);
 void stop_pwm(void);
 
 //solely notification purposes
